move temperature aware, soakable, and heating mode classes to core v1 package
