+++ src/main/java/com/github/thedeathlycow/thermoo/api/core/v1/package-info.java
@@ -2,4 +2,4 @@
  * Package for all things temperature related. Includes classes directly tied to temperature and temperature changes,
  * as well as sub-packages for effects and events.
  */
-package com.github.thedeathlycow.thermoo.api.temperature;
+package com.github.thedeathlycow.thermoo.api.core.v1;
