+++ src/main/java/com/github/thedeathlycow/thermoo/api/predicate/TemperatureLootCondition.java
@@ -1,6 +1,6 @@
 package com.github.thedeathlycow.thermoo.api.predicate;
 
-import com.github.thedeathlycow.thermoo.api.temperature.TemperatureAware;
+import com.github.thedeathlycow.thermoo.api.core.v1.TemperatureAware;
 import com.mojang.serialization.MapCodec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
 import net.minecraft.advancements.criterion.MinMaxBounds;
