+++ src/main/java/com/github/thedeathlycow/thermoo/api/temperature/event/EnvironmentTickContext.java
@@ -1,8 +1,8 @@
 package com.github.thedeathlycow.thermoo.api.temperature.event;
 
 import com.github.thedeathlycow.thermoo.api.environment.component.EnvironmentComponentTypes;
-import com.github.thedeathlycow.thermoo.api.temperature.Soakable;
-import com.github.thedeathlycow.thermoo.api.temperature.TemperatureAware;
+import com.github.thedeathlycow.thermoo.api.core.v1.Soakable;
+import com.github.thedeathlycow.thermoo.api.core.v1.TemperatureAware;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.component.DataComponentMap;
 import net.minecraft.server.level.ServerLevel;
