+++ src/main/java/com/github/thedeathlycow/thermoo/api/temperature/status/v2/tag/TemperatureStatusTags.java
@@ -1,7 +1,7 @@
 package com.github.thedeathlycow.thermoo.api.temperature.status.v2.tag;
 
 import com.github.thedeathlycow.thermoo.api.ThermooRegistryKeys;
-import com.github.thedeathlycow.thermoo.api.temperature.TemperatureAware;
+import com.github.thedeathlycow.thermoo.api.core.v1.TemperatureAware;
 import com.github.thedeathlycow.thermoo.api.temperature.status.v2.TemperatureStatus;
 import com.github.thedeathlycow.thermoo.impl.Thermoo;
 import net.minecraft.tags.TagKey;
