+++ src/main/java/com/github/thedeathlycow/thermoo/impl/environment/EnvironmentHeatingMode.java
@@ -1,7 +1,7 @@
 package com.github.thedeathlycow.thermoo.impl.environment;
 
-import com.github.thedeathlycow.thermoo.api.temperature.HeatingMode;
-import com.github.thedeathlycow.thermoo.api.temperature.TemperatureAware;
+import com.github.thedeathlycow.thermoo.api.core.v1.HeatingMode;
+import com.github.thedeathlycow.thermoo.api.core.v1.TemperatureAware;
 
 /**
  * A heating mode to apply the environment heat resistances. This is currently only used internally by {@link com.github.thedeathlycow.thermoo.api.environment.event.ServerPlayerEnvironmentTickEvents}
