+++ src/main/java/com/github/thedeathlycow/thermoo/impl/environment/EnvironmentTickContextImpl.java
@@ -1,7 +1,7 @@
 package com.github.thedeathlycow.thermoo.impl.environment;
 
-import com.github.thedeathlycow.thermoo.api.temperature.Soakable;
-import com.github.thedeathlycow.thermoo.api.temperature.TemperatureAware;
+import com.github.thedeathlycow.thermoo.api.core.v1.Soakable;
+import com.github.thedeathlycow.thermoo.api.core.v1.TemperatureAware;
 import com.github.thedeathlycow.thermoo.api.temperature.event.EnvironmentTickContext;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.component.DataComponentMap;
