+++ src/main/java/com/github/thedeathlycow/thermoo/mixin/common/EnvironmentAwareEntityMixin.java
@@ -2,9 +2,9 @@ package com.github.thedeathlycow.thermoo.mixin.common;
 
 import com.github.thedeathlycow.thermoo.api.ThermooAttributes;
 import com.github.thedeathlycow.thermoo.api.ThermooTags;
-import com.github.thedeathlycow.thermoo.api.temperature.HeatingMode;
-import com.github.thedeathlycow.thermoo.api.temperature.Soakable;
-import com.github.thedeathlycow.thermoo.api.temperature.TemperatureAware;
+import com.github.thedeathlycow.thermoo.api.core.v1.HeatingMode;
+import com.github.thedeathlycow.thermoo.api.core.v1.Soakable;
+import com.github.thedeathlycow.thermoo.api.core.v1.TemperatureAware;
 import com.github.thedeathlycow.thermoo.impl.component.ThermooComponents;
 import net.minecraft.core.Holder;
 import net.minecraft.util.Mth;
